requirements.txt
langchain
langchain-community
langchain-groq
ipywidgets
streamlit
python-dotenv